+++ server/src/parser/preprocessor.ts
@@ -1,9 +1,23 @@
 //import { URI } from 'vscode-languageserver';
 import { exec } from 'child_process';
-import { readFileSync } from 'fs';
-import { connection, preprocessedFilesCacheMap } from '../server';
+import { readFile, readFileSync } from 'fs';
+import { URI } from 'vscode-uri';
+import { connection, documents, preprocessedFilesCacheMap } from '../server';
 
 const rowsMap = new Map<string, number>();
+const defineMacrosMap = new Map();
+const macrosChecked = new Set();
+
+export function getDefineMacrosMap(): Map<string, any> {
+	return defineMacrosMap;
+}
+
+export function markFileToBeCheckedForMacroDefinitions(uri: string) {
+	macrosChecked.delete(uri);
+}
+
+const defineRegExp = new RegExp(/^\s*[#]define\s+([^(\n]+)/);
+
 
 export function runCommand(command: string) {
 	return new Promise( (resolve, reject) => {
@@ -167,8 +181,56 @@ function postProcessPreprocessedResult(unprocessedRowsMap: Map<string, number>) 
 				}	
 			}
 		}
+		if(isFileDefinesToBeProcess(filename)) {
+			mapMacroDefinitions(filename);
+		}
 	}
 	const elapsedTime = Date.now() - startTime;
 	connection.console.log(`Postprocessing row lines done in ${elapsedTime} ms`);
 }
 
+
+
+/**
+ * Determine if the file should be checked for define macros
+ * This preferably done once for library files and each time 
+ * for the currently opened file.
+ * 
+ * @param filename 
+ * @returns 
+ */
+function isFileDefinesToBeProcess(filename: string) {
+	return !macrosChecked?.has(filename);
+}
+
+function mapMacroDefinitions(filename: string) {
+	const uri = URI.parse(filename);
+	const fp = uri.fsPath;
+	readFile(fp, (err, data) => {
+		if(!err) {
+			let rowIdx = 0;
+			const text = data.toString();
+			const rows = text.split(/\r?\n/) ?? [];
+			rows.forEach((row, idx)=> {
+				const result = defineRegExp.exec(row);
+				if(result !== null && result?.length >= 2) {
+					let macroRows = 1;
+					if(row.includes('objFor')) {
+						console.error();
+					}
+					for(let nextRowIdx = idx; nextRowIdx<rows.length;nextRowIdx++) {
+						if(rows[nextRowIdx].endsWith('\\')) {
+							macroRows++;
+						} else {
+							break;
+						}
+					}
+					defineMacrosMap.set(result[1], {row: rowIdx, uri: fp, endLine: rowIdx+macroRows});
+				}
+				rowIdx++;
+			})
+		}
+		connection.console.info(`${uri.fsPath} scanned for macro definitions`);
+		macrosChecked.add(uri.fsPath); // Cache result, only refresh on file changes
+	});
+}